Updated functions and Styles
Fixed the maths to calculate the amount per pound. Made it so the big numbers have commas for readability. Removed irrelevant information and changed the reg_status to say registered or not instead of R or RM for user clarity. Changed the order of the page and the styles. Added some error handling to avoid errors returning when charities had little to no data. Added media queries so it fits on smaller screens and edited some styling.